requirements.txt
boto3
numpy
torch
torchvision
matplotlib
seaborn
scikit-learn
scipy
tensorflow-datasets
tensorflow-hub
tensorflowjs
imageio
mpld3
cython
opencv-python
pydot
